feat: extract Rooms tools into McpRoomsLib #include library + Ruff CI (#209)
First real module of the issue #209 modularization.

Modularization:
- Move the room tool IMPLEMENTATIONS (toolListRooms/toolGetRoom/toolCreateRoom/
  toolDeleteRoom/toolRenameRoom) from hubitat-mcp-server.groovy into
  libraries/mcp-rooms-lib.groovy (#include mcp.McpRoomsLib). Tool definitions,
  the hub_manage_rooms/hub_read_rooms gateway entries, and the executeTool
  dispatch cases stay in the app, so sandbox_lint's tool-count parser is
  unaffected and ToolRoomsSpec passes unchanged (the harness inlines the library).
- Register McpRoomsLib in getPackageLibraryRegistry() (hub_update_package leg).
- Rename the HPM bundle mcp-smoke-test.zip -> mcp-libraries.zip (a permanent
  multi-library bundle) carrying the Rooms library + the throwaway smoke canary;
  update packageManifest.json and tools/build-bundle.py.

Tests / e2e:
- mcp_install_bundle_hpm.sh now deletes + verifies EVERY bundle-managed library,
  so Rooms delivery via the bundle is genuinely proven (not masked by the
  pre-deploy user-library copy).
- e2e: assert McpRoomsLib is bundle-delivered; new self-cleaning
  create/get/rename/delete round-trip (device assigned) covering all five room
  tools live; Layer-6 room cleanup sweep; pre-existing-strand guard; harden
  test_list_libraries to require the populated library-API shape.
- IncludeResolverSpec resolves the real McpRoomsLib; BAT T509 updated.
- README manual-install now installs libraries before the parent app.

Housekeeping:
- Remove six .diff_*.txt working-artifacts accidentally committed in #243 + gitignore.

CI:
- Add a Ruff lint gate (ruff.toml + .github/workflows/ruff.yml, lint-only) and
  clear the pre-existing Python lint debt (F541/F401/E741/E402).
